feat: use types from monaco pkg
